+++ routes/shop.js
@@ -1,22 +1,9 @@
-const path = require('path');
-
 const express = require('express');
-
-const rootDir = require('../util/path');
-const adminData = require('../routes/admin');
 
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-	const products = adminData.products;
-	res.render('shop', {
-		prods: products,
-		pageTitle: 'Shop',
-		path: '/',
-		hasProducts: products.length > 0,
-		activeShop: true,
-        productCSS: true,
-	});
-});
+const productsController = require('../controllers/products');
+
+router.get('/', productsController.getProducts);
 
 module.exports = router;
